server/requirements.txt
# OCR Service Dependencies for NovaAgent
# Install: pip install -r server/requirements.txt

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# PDF processing
pymupdf==1.24.0
pdfplumber==0.11.0

# Image processing
Pillow==10.3.0

# OCR engines
pytesseract==0.3.13

# Optional: Better OCR (PaddleOCR)
# Uncomment to install PaddleOCR (requires more system dependencies)
# paddlepaddle==2.6.1
# paddleocr==2.8.1

# Utilities
pydantic==2.8.0
